+++ src/main/java/snw/rfm/api/events/HunterCatchPlayerEvent.java
@@ -30,7 +30,7 @@ public final class HunterCatchPlayerEvent extends Event implements Cancellable {
     private final int playerRemaining;
     private boolean isCancelled = false;
 
-    public HunterCatchPlayerEvent(Player catched, Player catcher, int playerRemaining) {
+    public HunterCatchPlayerEvent(Player catched, Player catcher, int playerRemaining, int coinEarned) {
         this.whoBeCatched = catched;
         this.catcher = catcher;
         this.playerRemaining = playerRemaining;
@@ -61,6 +61,18 @@ public final class HunterCatchPlayerEvent extends Event implements Cancellable {
      */
     public int getPlayerRemaining() {
         return playerRemaining;
+    }
+
+    /**
+     * 获取该玩家最终能获得的 B币 。
+     * <p>
+     * <i>因需要插件配置文件的配合，故 API 无法提供此方法的实现。</i>
+     *
+     * @param multiplier 决定获得的数据是否被处理过 (即是否受到插件配置 "coin_multiplier_on_be_catched" 的影响)
+     * @return 该玩家最终获得的 B币 数量。
+     */
+    public int getCoinEarned(boolean multiplier) {
+        throw new UnsupportedOperationException();
     }
 
     public static HandlerList getHandlerList() {